worker/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
playwright==1.40.0
dnspython==2.4.2
cryptography==41.0.7
aiohttp==3.9.1
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-mock==3.12.0
google-generativeai==0.8.3
